api user

+++ database/migrations/2019_04_16_040409_create_user_log.php
@@ -14,7 +14,7 @@ class CreateUserLog extends Migration
     public function up()
     {
         Schema::create('User_Log', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->Integer('id');
             $table->dateTime('last_login_date');
             $table->dateTime('last_job_apply_date');
         });
